+++ src/main/scala/io/github/scala_tessella/tessella/conversion/ConverterSVG.scala
@@ -4,7 +4,7 @@ package conversion
 import ConverterSVG.Description
 import SharedML.*
 import Geometry.Radian
-import GeometryBase.*
+import Geometry.*
 
 import scala.jdk.CollectionConverters.*
 import scala.xml.{Elem, Null, UnprefixedAttribute}
