+++ src/test/scala/io/github/scala_tessella/tessella/InteriorAngleSpec.scala
@@ -1,5 +1,6 @@
 package io.github.scala_tessella.tessella
 
+import Coordinates.*
 import Geometry.*
 import Geometry.Radian.TAU
 import RegularPolygon.Polygon
